requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
APScheduler==3.10.1
asttokens==2.2.1
async-timeout==4.0.2
asyncmy==0.2.5
attrs==23.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
cryptography==40.0.2
dnspython==2.2.1
docutils==0.19
ecdsa==0.18.0
email-validator==1.3.0
executing==1.2.0
fastapi==0.95.2
fastapi-base-tools==0.0.8
frozenlist==1.3.3
h11==0.9.0
httptools==0.1.2
icecream==2.1.3
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
jaeger-client==4.6.1
Jinja2==3.1.2
lockfile==0.12.2
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
opentracing==2.3.0
orjson==3.8.1
pandas==2.0.1
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.7.0
PyMySQL==1.0.3
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.5.4
requests==2.28.1
rsa==4.9
sentry-sdk==1.22.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadloop==1.0.2
thrift==0.16.0
tortoise-orm==0.19.3
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.22.0
watchfiles==0.18.0
yarl==1.9.2
